backend/requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pgvector==0.3.5

# Redis / Celery
celery==5.4.0
redis==5.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# OpenAI
openai==1.51.0

# Storage (Cloudflare R2 via S3 API)
boto3==1.35.36

# Utils
httpx==0.27.2
